test(cli): align target collection CI checks
