+++ crates/vize/src/commands/fmt/files.rs
@@ -8,6 +8,10 @@ use super::patterns::is_format_extension;
 const NODE_MODULES_DIR: &str = "node_modules";
 const VIZE_CACHE_DIR: &str = ".vize";
 
+#[cfg(test)]
+#[path = "files_tests.rs"]
+mod files_tests;
+
 #[allow(clippy::disallowed_types)]
 pub(crate) fn collect_files(
     patterns: &[impl AsRef<str>],
@@ -255,32 +259,6 @@ mod tests {
                 src.join("types.d.ts"),
             ]
         );
-    }
-
-    #[test]
-    fn collect_files_ignores_supported_extension_directories() {
-        let root = unique_case_dir("format-extension-directories");
-        let src = root.join("src");
-        let component_dir = src.join("Directory.vue");
-        let _ = fs::remove_dir_all(&root);
-        fs::create_dir_all(&component_dir).unwrap();
-        fs::write(src.join("App.vue"), "<template><div/></template>").unwrap();
-        fs::write(
-            component_dir.join("Nested.vue"),
-            "<template><div/></template>",
-        )
-        .unwrap();
-
-        let pattern = root.to_string_lossy().into_owned();
-        let glob_pattern = root.join("**/*.vue").to_string_lossy().into_owned();
-        let files_from_dir = collect_files(&[pattern], None);
-        let files_from_glob = collect_files(&[glob_pattern], None);
-        let _ = fs::remove_dir_all(&root);
-
-        let mut expected = vec![component_dir.join("Nested.vue"), src.join("App.vue")];
-        expected.sort();
-        assert_eq!(files_from_dir, expected);
-        assert_eq!(files_from_glob, expected);
     }
 
     #[test]
