+++ tests/tooling/cli-check-collection.test.ts
@@ -120,7 +120,7 @@ test("no-match text run prints the notice to stderr, leaves stdout empty, exits 
     assert.equal(result.stdout.trim(), "");
     assert.equal(
       result.stderr,
-      'No Vue, TypeScript, or JSX files found matching inputs: ["zzz.vue"]\n',
+      'No .vue, .ts, .tsx, .mts, .cts, .jsx, or .d.ts files found matching inputs: ["zzz.vue"]\n',
     );
   });
 });
