+++ tests/tooling/cli-check-contract.test.ts
@@ -105,7 +105,7 @@ test("vize check exits 0 when explicit inputs match no supported files", () => {
     assert.equal(result.stdout, "");
     assert.equal(
       result.stderr,
-      'No Vue, TypeScript, or JSX files found matching inputs: ["does-not-exist.vue"]\n',
+      'No .vue, .ts, .tsx, .mts, .cts, .jsx, or .d.ts files found matching inputs: ["does-not-exist.vue"]\n',
     );
   });
 });
@@ -115,6 +115,9 @@ test("vize check exits 0 in an empty project with no inputs", () => {
     const result = runCheck([], dir);
     assert.equal(result.status, 0, `${result.stdout}\n${result.stderr}`);
     assert.equal(result.stdout, "");
-    assert.equal(result.stderr, "No Vue, TypeScript, or JSX files found matching inputs: []\n");
+    assert.equal(
+      result.stderr,
+      "No .vue, .ts, .tsx, .mts, .cts, .jsx, or .d.ts files found matching inputs: []\n",
+    );
   });
 });
